Making the first askers in a locality editors #116
For now, it's set on 3 first askers, also added static messages to let
users know we need editors and that posting the question will make them
editors.
while working on the changes, I've improved the logic of when the "Ask
Question" button should appear and blocked users from asking questions
not in their locality.