Fixing #82 - search and refactoring entity passing
When views want to get a localized version of the page they need to add
the line `setattr(request, 'entity', entity)` before building the
context. The project's context processor then gets the entity and sets
the links and the intial value of the entity form.
TODO: create an entities/context_processor.py to handle it