Ensuring we have a default entity
We're late, so this is a quick & dirty way to ensure there's always an
entity in the context. The context processor is getting to big and IMO,
we should spilt it.
There's also some cleanup to d of old code that deals with the global
scope, which we now don't have, as we always have a default