#232 disable buttons when question is deleted
We don't want users to upvote/flag/delete/edit questions when they are
deleted, so the controls are now disabled in a deleted question page.
However, the models still support all the actions if the user is feeling
smart and hacks the request, but it's not such a big deal.