#258 show locality when asking a question
The 'entity' variable is not passed to the context processor when
rendering the post_question form. It is being set in the context
dictionary, but the context processor does not see it - it just sees the
request, which does not contain the entity because it does not appear in
the URL for post_question.
Instead, get the entity from the user's locality if they are logged in
(otherwise they would not have been able to post a question in the first
place).