Fix edit profile locality required message
The "this field is required" message was being displayed for the locality in
the "edit profile" form even if the user already had a locality set (so that
the field is pre-set and disabled).
To fix this, I set the field to required=False again but validated it in the
new clean_locality method.