Show candidate list in locality page when clicked
The default view is to show the mayorship candidates, but now when a
candidate list is clicked, its candidates will be shown instead. To go
to the page of the candidate list, click the title of the candidates
section in the locality page after clicking the candidate list icon in
the navigation bar.
To go back to view the mayorship candidates, click the yellow note which
is displayed above the candidate lists if a list is selected.