Replace profile.locality by a many-to-many field: entities
Added through a Membership class that includes an is_editor field instead of
profile.is_editor, because a user can be an editor in one entity and not an
editor in another.

This means there are currently two many-to-many relationships between profile
and entity: through Membership and through Candidate. This should be fixed in
the future (see #454).

Also added an entity parameter to profile.is_candidate and added a method
profile.is_editor which also has an entity parameter.

Work is not complete yet. Still need to replace occurrences of 'locality',
is_candidate and is_editor.
Removed is_editor from Admin interface for User, but it still needs to be added
as a field in the profile inline: this can be done by nested inlines (currently
not supported in our version of Django) or eliminating Profile altogether and
extending User instead (as done in Open Knesset).